Split business logic into separate hook
